Add a url_hash column for faster SQL lookups with large tables
SHA1 was chosen as the hash algorithm since it is supported by mysql for the upgrade function.
Otherwise the upgrade would need to be done in PHP on each row which would be much slower on large tables.

Only the first 8 characters of the hash are stored to minimize storage and index requirements on the database.
Even though SHA1 is considered cryptographially inscure at this point and we are using only the first 8 characters making the possiblitly of collisions even higher.
That is okay since we are only using it to index the urls in the database for faster lookups.
Even with many collisions it is still substantially faster to lookup since it will still match on the indexed hashes and then only need to filter down to the matching url (which we still do in the queries).

Fixes #3793